ci: move service to another folder
